+++ .gitignore
@@ -1,9 +1,7 @@
-VideoSlim/.idea/
-VideoSlim/log.txt
-VideoSlim/store
-VideoSlim/old_atemp.mp4
-VideoSlim/old_atemp.wav
-VideoSlim/old_vtemp.mp4
+.idea/
+log.txt
+store
+/output/
 /.idea
 /.vscode
 /.venv
