--- VideoSlim/src/meta.py
@@ -1,21 +1,0 @@
-"""
-VideoSlim 应用程序常量定义
-"""
-
-# 应用程序版本号
-VERSION = "v1.8"
-
-# 检查更新的 GitHub API URL
-CHECK_UPDATE_URL = "https://api.github.com/repos/mainite/VideoSlim/releases"
-
-# 支持的视频文件扩展名列表
-SUPPORTED_VIDEO_EXTENSIONS = [".mp4", ".mkv", ".mov", ".avi"]
-
-# 配置文件路径
-CONFIG_FILE_PATH = "config.json"
-
-# 存储文件路径
-STORE_PATH = "store"
-
-# 临时文件路径列表
-TEMP_FILES = ["./pre_temp.mp4"]
