+++ pyproject.toml
@@ -1,0 +1,12 @@
+[project]
+name = "videoslim"
+version = "2.0.0"
+description = "简洁易用的 Windows 视频压缩工具"
+readme = "README.md"
+requires-python = ">=3.12"
+dependencies = [
+    "pydantic>=2.12.5",
+    "pymediainfo>=7.0.1",
+    "requests>=2.32.5",
+    "windnd>=1.0.7",
+]
