+++ src/model/__init__.py
@@ -1,0 +1,11 @@
+"""
+VideoSlim - 数据模型模块
+
+定义应用程序使用的所有数据模型和接口，包括视频文件、任务、消息、配置和存储等。
+"""
+
+# 导入所有模型模块，使它们可以通过src.model直接访问
+from . import config
+from . import message
+from . import store
+from . import video
